Wymuszenie komunikacji z bazą danych w kodowaniu utf8; zmiana typu zmiennej ceny, aby obsługiwała części złotych; wprowadzenie projekcji do repozytorium rest-owego Uzytkownik w celu wyświetlania informacji o użytkowniku, który dodał daną ofertę w podglądzie oferty; dodanie dodatkowego rekordu do bazy danych do tabeli oferty; zmiana wersji na v0.0.3 w kodzie.

+++ src/main/java/com/magicalsolutions/fastcourier/Projection/OfertaProjection.java
@@ -51,14 +51,14 @@ public interface OfertaProjection {
      * @author Piotr Podbielski
      * @return cena minimalna
      */
-    Integer getCenaMinimalna();
+    Double getCenaMinimalna();
 
     /**
      * Funkcja zwracająca cenę maksymalną
      * @author Piotr Podbielski
      * @return cena maksymalna
      */
-    Integer getCenaMaksymalna();
+    Double getCenaMaksymalna();
 
     /**
      * Funkcja zwracająca miasto początkowe
